# gate/target/generated-sources/openapi/src/main/java/ru/maximenko/data/api/StudentGateApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package ru.maximenko.data.api;

import ru.maximenko.data.model.StudentGateCreateRequest;
import ru.maximenko.data.model.StudentGateResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-25T17:51:18.249279700+03:00[Europe/Moscow]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "studentGate", description = "the studentGate API")
public interface StudentGateApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    public static final String PATH_CREATE_STUDENT = "/api/students";
    /**
     * POST /api/students : Проксировать создание студента во внутренний сервис
     * Принимает запрос от клиента и перенаправляет его в DATA SERVICE.
     *
     * @param studentGateCreateRequest  (required)
     * @return Студент создан (ответ от DATA SERVICE) (status code 201)
     */
    @Operation(
        operationId = "createStudent",
        summary = "Проксировать создание студента во внутренний сервис",
        description = "Принимает запрос от клиента и перенаправляет его в DATA SERVICE.",
        tags = { "studentGate" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Студент создан (ответ от DATA SERVICE)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = StudentGateResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = StudentGateApi.PATH_CREATE_STUDENT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<StudentGateResponse> createStudent(
        @Parameter(name = "StudentGateCreateRequest", description = "", required = true) @Valid @RequestBody StudentGateCreateRequest studentGateCreateRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"passport\" : \"1234 567890\", \"fullName\" : \"Иванов Иван Иванович\", \"id\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_STUDENT_BY_ID = "/api/students/{studentId}";
    /**
     * GET /api/students/{studentId} : Получить данные студента по ID
     * Проксирует запрос на получение студента во внутренний DATA SERVICE.
     *
     * @param studentId ID студента (required)
     * @return Данные студента (status code 200)
     *         or Студент не найден (status code 404)
     *         or Внутренняя ошибка сервера (status code 500)
     */
    @Operation(
        operationId = "getStudentById",
        summary = "Получить данные студента по ID",
        description = "Проксирует запрос на получение студента во внутренний DATA SERVICE.",
        tags = { "studentGate" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Данные студента", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = StudentGateResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Студент не найден"),
            @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервера")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = StudentGateApi.PATH_GET_STUDENT_BY_ID,
        produces = { "application/json" }
    )
    
    default ResponseEntity<StudentGateResponse> getStudentById(
        @Parameter(name = "studentId", description = "ID студента", required = true, in = ParameterIn.PATH) @PathVariable("studentId") Long studentId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"passport\" : \"1234 567890\", \"fullName\" : \"Иванов Иван Иванович\", \"id\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}